+++ blueprints/stacks_bp.py
@@ -1,9 +1,9 @@
 from flask import Blueprint, request
-# from flask_jwt_extended import jwt_required, get_jwt_identity
+from flask_jwt_extended import jwt_required
 from models.stack import Stack, StackSchema
 from setup import db
 from blueprints.stack_tools_bp import stack_tools_bp
-# from auth import authorize
+from auth import authorize
 
 
 stacks_bp = Blueprint("stacks", __name__, url_prefix="/stacks")
@@ -12,18 +12,19 @@ stacks_bp.register_blueprint(stack_tools_bp)
 
 # Get all stacks
 @stacks_bp.route("/")
-# @jwt_required()
 def all_stacks():
     
     stmt = db.select(Stack).order_by("id")
     stacks = db.session.scalars(stmt).all()
-    return StackSchema(many=True).dump(stacks)
+    return StackSchema(many=True, exclude=["stack_tools"]).dump(stacks)
 
 
 # Create a stack
 @stacks_bp.route("/", methods=["POST"])
-# @jwt_required()
+@jwt_required()
 def create_stack():
+    
+    authorize()
     
     stack_info = StackSchema(exclude=["id"]).load(request.json)
     
@@ -40,14 +41,12 @@ def create_stack():
 
 # Get a single stack
 @stacks_bp.route("/<int:id>")
-# @jwt_required()
 def one_stack(id):
     
     stmt = db.select(Stack).filter_by(id = id)
     stack = db.session.scalar(stmt)
     
     if stack:
-        # authorize(stack.user_id)
         return StackSchema().dump(stack)
     
     return {"error": "Stack not found"}, 404
@@ -55,8 +54,10 @@ def one_stack(id):
 
 # Update a single stack
 @stacks_bp.route("/<int:id>", methods=["PUT", "PATCH"])
-# @jwt_required()
+@jwt_required()
 def update_stack(id):
+    
+    authorize()
     
     stack_info = StackSchema(exclude=["id"]).load(request.json)
     
@@ -64,7 +65,6 @@ def update_stack(id):
     stack = db.session.scalar(stmt)
     
     if stack:
-        # authorize(stack.user_id)
         stack.name = stack_info.get("name", stack.name)
         stack.description = stack_info.get("description", stack.description)
         db.session.commit()
@@ -75,14 +75,15 @@ def update_stack(id):
 
 # Delete a single stack
 @stacks_bp.route("/<int:id>", methods=["DELETE"])
-# @jwt_required()
+@jwt_required()
 def delete_stack(id):
+    
+    authorize()
     
     stmt = db.select(Stack).filter_by(id = id)
     stack = db.session.scalar(stmt)
     
     if stack:
-        # authorize(stack.user_id)
         db.session.delete(stack)
         db.session.commit()
         return {"status": f"{stack.name} has been deleted"}, 200
