+++ blueprints/users_bp.py
@@ -1,25 +1,28 @@
 from flask import Blueprint, request
-# from auth import authorize
+from auth import authorize
 from flask_jwt_extended import create_access_token, jwt_required
 from datetime import timedelta
 from models.user import User, UserSchema
 from blueprints.user_tools_bp import user_tools_bp
+# from blueprints.user_plans_bp import user_plans_bp
 from setup import bcrypt, db
 from sqlalchemy import exc
 
 
 users_bp = Blueprint("users", __name__, url_prefix="/users")
 users_bp.register_blueprint(user_tools_bp)
+# users_bp.register_blueprint(user_plans_bp)
 
 # Get all users
 @users_bp.route("/")
-# @jwt_required()
+@jwt_required()
 def all_users():
-    # authorize()
+    
+    authorize()
     
     stmt = db.select(User)
     users = db.session.scalars(stmt).all()
-    return UserSchema(many=True, exclude=["password"]).dump(users)
+    return UserSchema(many=True, exclude=["password", "user_tools"]).dump(users)
 
 
 # Register a new user
@@ -31,7 +34,7 @@ def register():
         user = User(
             email = user_info["email"],
             password = bcrypt.generate_password_hash(user_info["password"]).decode("utf8"),
-            name = user_info.get("name", "anonymous")
+            name = user_info.get("id", "name", "anonymous")
         )
         
         db.session.add(user)
@@ -51,22 +54,22 @@ def login():
     user = db.session.scalar(stmt)
     
     if user and bcrypt.check_password_hash(user.password, user_info["password"]):
-        token = create_access_token(identity=user.id, expires_delta=timedelta(hours=2))
-        return {"token": token, "user": UserSchema(only=["email", "name"]).dump(user)}
+        token = create_access_token(identity=user.id, expires_delta=timedelta(hours=100)) # To adjust timeout later
+        return {"token": token, "user": UserSchema(only=["id", "email", "name"]).dump(user)}
     else:
         return {"error": "Invalid email or password"}, 401
     
 
 # Get a single user
 @users_bp.route("/<int:id>")
-# @jwt_required()
+@jwt_required()
 def one_user(id):
     
     stmt = db.select(User).filter_by(id = id)
     user = db.session.scalar(stmt)
     
     if user:
-        # authorize(user.user_id)
+        authorize(user.user_id)
         return UserSchema(only=["id", "email", "name"]).dump(user)
     
     return {"error": "User not found"}, 404    
@@ -74,7 +77,7 @@ def one_user(id):
     
 # Update a single user
 @users_bp.route("/<int:id>", methods=["PUT", "PATCH"])
-# @jwt_required()
+@jwt_required()
 def update_user(id):
     
     user_info = UserSchema(exclude=["id"]).load(request.json)
@@ -83,7 +86,7 @@ def update_user(id):
     user = db.session.scalar(stmt)
     
     if user:
-        # authorize(user.user_id)
+        authorize(user.user_id)
         user.name = user_info.get("name", user.name)
         user.email = user_info.get("email", user.email)
         user.password = user_info.get("password", user.password)
@@ -95,14 +98,14 @@ def update_user(id):
 
 # Delete a single user
 @users_bp.route("/<int:id>", methods=["DELETE"])
-# @jwt_required()
+@jwt_required()
 def delete_user(id):
     
     stmt = db.select(User).filter_by(id = id)
     user = db.session.scalar(stmt)
     
     if user:
-        # authorize(user.user_id)
+        authorize(user.user_id)
         db.session.delete(user)
         db.session.commit()
         return {"status": f"{user.name} has been deleted"}, 200
